refactor(cli): extract hook bridge setup (Phase C, the big one)
The last big cut of the createNewSession elephant: the 386-line
if (hookServer) {...} block that wired all 7 Claude Code hook events
into our PTY's MessageAPI.

Three tangled concerns moved together because they share per-session
locks (claudeSessionId, mainSessionEnded, hasSiblingInDir):
  1. Session filtering: drop subagent and sibling-daemon events so they
     don't hijack our status/questions/transcript watching.
  2. Transcript discovery via hooks: most events carry transcript_path,
     letting us skip the 2s mtime poll; a restart (/clear /compact
     /resume) tears down the old watcher and starts a fresh one.
  3. Auto-approve gate: PermissionRequest is intercepted before creating
     a Question; approve/deny inject '1'/'3' into the PTY; escalate path
     falls back to the user, or default-denies in a subagent context to
     avoid hanging with no one to answer.

- new cli/session-phases/hook-bridge-setup.ts (477 lines):
  setupHookBridge(deps, args) registers 7 hookServer listeners
  (SessionStart, PreToolUse, PostToolUse, Notification, PermissionRequest,
  Stop, SessionEnd) and returns void. Runs once per session at
  createNewSession time, only when a hookServer is configured.
- deps bundle: sessionRegistry, sessionStore, liveSessionsRegistry,
  transcriptWatchers, transcriptFallbackTimers, autoApproveService,
  currentPort getter. args bundle: hookServer, sessionId,
  workingDirectory, messageApi, sendAndRecord.
- cli.ts: replaces the 386-line block with a 12-line call. Also deletes
  the now-unused startTranscriptWatcher thin wrapper that was the last
  cli.ts caller of the pre-#352 helper.
- tests: 4 using a RecordingHookServer (records .on() registrations
  without starting the real Bun.serve HTTP listener). Verifies all 7
  listeners register, the function returns void, no throw with null
  auto-approve, and registrations are idempotent across sessions.
  Callback runtime behavior is byte-for-byte preserved from the inline
  version and is covered by tests/integration/run-tests.sh.
- cli.ts: 2247 -> 1855 (-392)
